fix bug

+++ ExpenseMaster/Controllers/ExpenseController.cs
@@ -7,9 +7,9 @@ using Microsoft.AspNetCore.Mvc;
 namespace ExpenseMaster.Controllers
 {
     [ApiController]
-    [Route("public")]
+    [Route("public/expenses")]
     [AllowAnonymous]
-    public class ExpenseController : Controller
+    public class ExpenseController : ControllerBase
     {
         private readonly IExpenseService _expenseService;
 
@@ -22,7 +22,6 @@ namespace ExpenseMaster.Controllers
         public async Task<ActionResult<IEnumerable<Expense>>> GetExpenses()
         {
             var expenses = await _expenseService.GetAllExpenses();
-
             return Ok(expenses);
         }
 
@@ -30,7 +29,10 @@ namespace ExpenseMaster.Controllers
         public async Task<ActionResult<Expense>> GetExpenseById(int id)
         {
             var expense = await _expenseService.GetExpenseById(id);
-
+            if (expense == null)
+            {
+                return NotFound();
+            }
             return Ok(expense);
         }
 
@@ -38,43 +40,51 @@ namespace ExpenseMaster.Controllers
         public async Task<ActionResult<Expense>> CreateExpense(CreateExpenseDto createExpenseDto)
         {
             var expense = await _expenseService.CreateExpense(createExpenseDto);
-
             return Ok(expense);
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<Expense>> UpdateExpense(Expense expense)
+        public async Task<ActionResult<Expense>> UpdateExpense(int id, Expense expense)
         {
-            var existingExpense = await _expenseService.GetExpenseById(expense.Id);
+            if (id != expense.Id)
+            {
+                return BadRequest();
+            }
 
-            var expenseToUpdate = _expenseService.UpdateExpense(existingExpense);
+            var existingExpense = await _expenseService.GetExpenseById(id);
+            if (existingExpense == null)
+            {
+                return NotFound();
+            }
 
-            return Ok(expenseToUpdate);
+            var updatedExpense = await _expenseService.UpdateExpense(expense);
+            return Ok(updatedExpense);
         }
 
         [HttpDelete("{id}")]
         public async Task<ActionResult<int>> DeleteExpense(int id)
         {
             var existingExpense = await _expenseService.GetExpenseById(id);
+            if (existingExpense == null)
+            {
+                return NotFound();
+            }
 
             await _expenseService.DeleteExpense(existingExpense);
-
             return Ok(existingExpense.Id);
         }
 
-        [HttpGet("expenses/category/{categoryId}")]
+        [HttpGet("category/{categoryId}")]
         public async Task<ActionResult<IEnumerable<Expense>>> GetExpensesByCategory(int categoryId)
         {
             var expenses = await _expenseService.GetExpensesByCategory(categoryId);
-
             return Ok(expenses);
         }
 
-        [HttpGet("expenses/user/{userId}/total")]
+        [HttpGet("user/{userId}/total")]
         public async Task<ActionResult<decimal>> CalculateTotalExpenseByUserId(int userId)
         {
             var totalExpenses = await _expenseService.CalculateTotalExpensesByUserId(userId);
-
             return Ok(totalExpenses);
         }
     }
